requirements.txt
pytz==2018.5
cx-Oracle==6.4.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-camel-case==0.2.0
tokenlib==2.0.0